requirements.txt
tweepy
openai
schedule
python-dotenv
requests
